Use `docker compose` instead of `docker-compose` (#1452)
* Use `docker compose` instead of `docker-compose`

We also no longer need to explicitly check the `docker-compose` version -
checking the general `docker` version should suffice.

* Update docs/fides/docs/tutorial/add.md

* move the docker check to the top of the nox startup

* update check logic

* fix a bad change

* account for another possible version format

* add another check to see if the docker cli is available

* update a CI nox session to not recursively call nox

* fix static_checks
