+++ docs/fides/docs/tutorial/add.md
@@ -44,7 +44,7 @@ See our [Configuration](../installation/configuration.md) guide for more informa
 
 ## Run Fidesctl via Docker
 
-Now that the dependency is included in the project and the configuration is in place, the fidesctl server needs to be told to run. The app uses `docker-compose` to orchestrate resources, so include `fidesctl` as a service by adding the following configuration after the database service:
+Now that the dependency is included in the project and the configuration is in place, the fides server needs to be told to run. The app uses `docker compose` to orchestrate resources, so include `fides` as a service by adding the following configuration after the database service:
 
 ```yaml
 fidesctl:
@@ -64,7 +64,7 @@ fidesctl:
     FIDESCTL__DATABASE__USER: "postgres"
 ```
 
-> See [the fidesctl installation guide](../installation/overview.md) for a more detailed fidesctl server setup walkthrough, and [the `docker-compose` documentation](https://docs.docker.com/compose/compose-file/) for an explanation of the above configuration options.
+> See [the fidesctl installation guide](../installation/overview.md) for a more detailed fidesctl server setup walkthrough, and [the `docker compose` documentation](https://docs.docker.com/compose/compose-file/) for an explanation of the above configuration options.
 
 ## Check Your Progress
