+++ noxfiles/ci_nox.py
@@ -16,7 +16,6 @@ from constants_nox import (
 from run_infrastructure import OPS_TEST_DIR, run_infrastructure
 from utils_nox import install_requirements
 
-RUN_STATIC_ANALYSIS = (*RUN_NO_DEPS, "nox", "-s")
 
 # Static Checks
 @nox.session()
@@ -99,10 +98,7 @@ def check_install(session: nox.Session) -> None:
 @nox.session()
 def check_fides_annotations(session: nox.Session) -> None:
     """Run a fidesctl evaluation."""
-    if session.posargs == ["docker"]:
-        run_command = (*RUN_STATIC_ANALYSIS, "check_fides_annotations")
-    else:
-        run_command = ("fides", "--local", *(WITH_TEST_CONFIG), "evaluate")
+    run_command = (*RUN_NO_DEPS, "fides", "--local", *(WITH_TEST_CONFIG), "evaluate")
     session.run(*run_command, external=True)
 
 
@@ -139,7 +135,8 @@ def pytest_ctl(session: nox.Session, mark: str) -> None:
     session.notify("teardown")
     if mark == "external":
         start_command = (
-            "docker-compose",
+            "docker",
+            "compose",
             "-f",
             COMPOSE_FILE,
             "-f",
@@ -150,7 +147,8 @@ def pytest_ctl(session: nox.Session, mark: str) -> None:
         )
         session.run(*start_command, external=True)
         run_command = (
-            "docker-compose",
+            "docker",
+            "compose",
             "run",
             "-e",
             "SNOWFLAKE_FIDESCTL_PASSWORD",
@@ -255,7 +253,8 @@ def pytest_ops(session: nox.Session, mark: str) -> None:
         #
         session.run(*START_APP_WITH_EXTERNAL_POSTGRES, external=True)
         run_command = (
-            "docker-compose",
+            "docker",
+            "compose",
             "run",
             "-e",
             "ANALYTICS_OPT_OUT",
